Update app.R
Modified the server code to create datasetInput using eventReactive() function  that creates a data frame for predictor variables to be updated in the event of the hitting the submit button; as such there is not need to create a csv file for test dataset. The reactive code in both, the text box and table also get triggered upon event.